Separate keyword and value
In this scenario I want to separate two are not the same keyword
such as

Library    Selenium2Library

In this case, `Library' and `Selenium2Library' are not the same keyword.

In this commit, I changed from jasmine test suite to `atom-grammar-test'.
for easy to read and can be use in issue when found a bug.